# api/api.yml
openapi: 3.0.2
info:
  version: 1.0.0
  title: Issues Taiga
  description: "This is the Open API documentation for the REST API of our beloved application **Issues** deployed at <https://falta-link/>. <br>All operations are executable. All of the  operations requires authentication, you must **Authorize** your request by providing the token value you got when you created the tweet."
servers:
  # Added by API Auto Mocking Plugin
  - url: 'http://3.235.230.222:8080/'
paths:
...
  /issues/{issueId}:
    get:
      tags:
      - issues
      summary: Finds an Issue by its ID
      description: Returns a single Issue
      operationId: getIssueById
      parameters:
      - name: issueId
        in: path
        description: ID of the Tweet to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
          200:
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Issue'
          404:
            description: 'Error: Not Found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error404'
          403:
            description: 'Error: Forbidden'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error403'
    put:
      tags:
        - issues
      summary: Update an existing Issue with the new parameters passed
      description: Update an existing Issue by Id
      operationId: updateIssue
      parameters:
      - name: issueId
        in: path
        description: ID of the Issue to return
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update an existent Issue in the store
        content:
          application/json:
            schema:
              properties:
                subject:
                  type: string
                  minLength: 4
                  example: 'My subject'
                description:
                  type: string
                  minLength: 4
                  maxLength: 280
                  example: 'My description'
                typeIssue:
                  type: integer
                  minimum: 0
                  maximum: 3
                  example: 0
                severityIssue:
                  type: integer
                  minimum: 0
                  maximum: 4
                  example: 0
                priorityIssue:
                  type: integer
                  minimum: 0
                  maximum: 2
                  example: 0
                statusIssue:
                  type: integer
                  minimum: 0
                  maximum: 6
                  example: 0
                due_date:
                  type: string
                  format: date-time
                blocked:
                  type: boolean
                  example: false
                blocked_reason:
                  type: string
                  example: 'blocked reaon'
                user_id:
                  type: integer
                  example: 1
        required: true
      responses:
        200:
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Issue'
        404:
            description: 'Error: Not Found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error404'
      security:
      - ApiKeyAuth: []
    delete:
      tags:
      - issues
      summary: Deletes an Issue associated with the ID passed as parameter, if the ID does not exist, don't do nothing
      operationId: destroy
      parameters:
      - name: issueId
        in: path
        description: Issue id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        204:
          description: No content
          content:
            application/json: {}
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
      security:
      - ApiKeyAuth: []
  /issues/bulkForm:
    post:
      tags:
      - issues/bulk
      summary: Creates multiple new Issues
      operationId: bulkCreate
      requestBody:
        description: Provide the necessary fields to generate as Issues as you want
        content:
          application/json:
            schema:
              type: object
              properties:
                issues:
                  type: array
                  items:
                    type: object
                    properties:
                      subject:
                        type: string
                        minLength: 4
                        example: "Me mareo"
        required: true
      responses:
          200:
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Issue'
          404:
            description: 'Error: Not Found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error404'
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
    Issue:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        subject:
          type: string
          minLength: 4
          example: 'My subject'
        description:
          type: string
          example: 'My description'
        due_date:
          type: string
          format: date-time
        blocked:
          type: boolean
        blocked_reason:
          type: string
          example: 'blocked reaon'
        watchers:
          type: string
          example: 'Member X'
        typeIssue:
          type: integer
          minimum: 0
          maximum: 3
          example: 0
        severityIssue:
          type: integer
          minimum: 0
          maximum: 4
          example: 0
        priorityIssue:
          type: integer
          minimum: 0
          maximum: 2
          example: 0
        statusIssue:
          type: integer
          minimum: 0
          maximum: 6
          example: 0
        user_id:
          type: integer
          example: 1
        createdBy:
          type: integer
          example: 1
        _links:
          type: object
          properties:
            self:
              type: string
              format: url
              example: 'https://falta-link/issues/2/edit'
...
    ErrorResult:
      type: object
      properties:
        message:
          type: string
          example: 'Content is too long (maximum is 280 characters)'
      required:
        - message
  examples:
...
    error401:
      value:
        message: "You provided no token"
    error403:
      value:
        message: "You provided an invalid token"
    error404:
      value:
        message: "There is no issue with 'id'=8"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header